Add RangeInclusiveMap type
Uses RangeInclusive instead of Range as its keys.

This requires a bit of fiddling around with extra trait bounds
on the keys, because ranges can interact with each other by having
one range's start exactly one away from another range's end.

This requires use of the unstable Step trait (or some equivalent,
which I've introduced here) _and_ the ability to know the max
value of the key space, which I'm achieving using `num` for now.

The latter may be possible to relax with some different trickery
around the order of ranges we're comparing, and a different way
of short-circuiting some loops. I'm going to experiment with that
before committing to bringing in any extra dependency.