Allow specifying Step functions via separate param
Because `std::iter::Step` is not stable, I have my own `StepLite` trait.
But that means other people would need to implement `StepLite` for
foreign traits, so I've provided a bit of a hack to get around that:
providing the relevant functions via a separate type parameter,
with associated free functions.