Fix coalescing of contiguous ranges
Fixes https://github.com/jeffparsons/rangemap/issues/24

I'm not super-happy with this fix; even though it does address the
immediate problem, it also highlights how the complexity of this code
creates habitat for serious bugs.

What I really want is a cursor API for BTreeMap, because that would
make this entire crate nearly trivial. But I think I might be able
to rewrite a lot of it "pretending" that I have a cursor API, in a
way that would be a lot less error-prone, if slightly less performant.
So I'll give that a try when I find some time.