Fix a performance regression in 0.1.9.
This performance issue stems from the fix for #24 in d1999f4.
Previously, when the code was looking for "touching" ranges, it
applies the filter() block to at most one range (from next_back()).
But with the change in d1999f4, if none of the ranges pass the
filter, the code now potentially looks at a large number of ranges,
making insert() potentially quadratic in the map size (I think).

This patch fixes the issue by applying the "filter" to
".rev().take(2)" -- at most the two final ranges that could match.
This way, neither of the last two ranges are touching the
just-inserted range, then no more ranges are considered.

(I have verified that the tests and fuzzers still pass with this
change, but I do not know the code well enough to promise it is
correct.)