Improve use of dense map in tests
Now the sparse and dense maps are both converted to Vecs
(items in dense map are coalesced by its new iterator)
and compared. We'll get better messages out of test assertions
this way, and it all feels a little bit neater.

I've done this because I'm planning to use the dense map
in more tests.