Allow the `overlapping()` function to take ranges by value
Instead of forcing the user of the API to pass a range by reference,
which causes the resulting iterator to have a lifetime which depends on
this reference, the `Borrow` trait is used to give the user a choice to
use either a reference or pass the range by value. The lifetime of the
iterator accordingly depends on which of the two is chosen.